meta-isar: remove enable-fsck service for some qemu targets
enable-fsck service is designed to be run once at first boot in order
to make changes in fstab related to fsck flag on real partitions.

For qemu targets used mostly for CI it's not essential, but
leads to some problems e.g. long boot time and/or script failure
due to out-of-space error on / or /boot.

So, the service can be removed from default packages list
for these qemu targets

Signed-off-by: Uladzimir Bely <ubely@ilbers.de>